+++ vsd/src/downloader.rs
@@ -1,7 +1,7 @@
 use crate::{
     commands::Quality,
     merger::Merger,
-    playlist::{KeyMethod, MediaPlaylist, MediaType, PlaylistType, Range, Segment},
+    playlist::{KeyMethod, MasterPlaylist, MediaPlaylist, MediaType, PlaylistType, Range, Segment},
     update, utils,
 };
 use anyhow::{anyhow, bail, Result};
@@ -260,6 +260,82 @@ pub fn process_playlist(
                 };
                 crate::hls::push_segments(&m3u8, &mut media_playlist);
                 Ok((vec![media_playlist], vec![]))
+            }
+            Err(x) => bail!(
+                "couldn't parse response as hls playlist (failed with {}).\n\n{}\n\n{}",
+                x,
+                meta.url,
+                meta.text
+            ),
+        },
+        _ => bail!("couldn't determine playlist type, only DASH and HLS playlists are supported."),
+    }
+}
+
+pub fn parse_playlist(
+    base_url: Option<Url>,
+    client: &Client,
+    meta: &InputMetadata,
+) -> Result<MasterPlaylist> {
+    match meta.pl_type {
+        Some(PlaylistType::Dash) => {
+            let mpd = dash_mpd::parse(&meta.text).map_err(|x| {
+                anyhow!(
+                    "couldn't parse response as dash playlist (failed with {}).\n\n{}",
+                    x,
+                    meta.text
+                )
+            })?;
+            let mut playlist = crate::dash::parse_as_master(&mpd, meta.url.as_ref());
+
+            for stream in playlist.streams.iter_mut() {
+                crate::dash::push_segments(
+                    &mpd,
+                    stream,
+                    base_url.as_ref().unwrap_or(&meta.url).as_str(),
+                )?;
+                stream.uri = meta.url.as_ref().to_owned();
+            }
+
+            Ok(playlist)
+        }
+        Some(PlaylistType::Hls) => match m3u8_rs::parse_playlist_res(meta.text.as_bytes()) {
+            Ok(m3u8_rs::Playlist::MasterPlaylist(m3u8)) => {
+                let mut playlist = crate::hls::parse_as_master(&m3u8, meta.url.as_ref());
+
+                for stream in playlist.streams.iter_mut() {
+                    stream.uri = base_url
+                        .as_ref()
+                        .unwrap_or(&meta.url)
+                        .join(&stream.uri)?
+                        .to_string();
+                    let response = client.get(&stream.uri).send()?;
+                    let text = response.text()?;
+                    let media_playlist = m3u8_rs::parse_media_playlist_res(text.as_bytes())
+                        .map_err(|x| {
+                            anyhow!(
+                                "couldn't parse response as hls playlist (failed with {}).\n\n{}\n\n{}",
+                                x,
+                                stream.uri,
+                                text
+                            )
+                        })?;
+                    crate::hls::push_segments(&media_playlist, stream);
+                }
+
+                Ok(playlist)
+            }
+            Ok(m3u8_rs::Playlist::MediaPlaylist(m3u8)) => {
+                let mut media_playlist = crate::playlist::MediaPlaylist {
+                    uri: meta.url.as_ref().to_owned(),
+                    ..Default::default()
+                };
+                crate::hls::push_segments(&m3u8, &mut media_playlist);
+                Ok(MasterPlaylist {
+                    playlist_type: PlaylistType::Hls,
+                    streams: vec![media_playlist],
+                    uri: meta.url.as_ref().to_owned(),
+                })
             }
             Err(x) => bail!(
                 "couldn't parse response as hls playlist (failed with {}).\n\n{}\n\n{}",