+++ vsd/src/utils.rs
@@ -2,6 +2,7 @@ use aes::cipher::{block_padding::Pkcs7, BlockDecryptMut, KeyIvInit};
 use anyhow::{anyhow, bail, Result};
 use base64::Engine;
 use regex::Regex;
+use std::collections::HashSet;
 
 type Aes128CbcDec = cbc::Decryptor<aes::Aes128>;
 
@@ -43,19 +44,25 @@ pub(super) fn format_download_bytes(downloaded: usize, total: usize) -> String {
 }
 
 pub(super) fn scrape_playlist_links(text: &str) -> Vec<String> {
-    let re = Regex::new(r"(https|ftp|http)://([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-]\.(m3u8|m3u|mpd))").unwrap();
+    // let re = Regex::new(r"(https|ftp|http)://([\w_-]+(?:(?:\.[\w_-]+)+))([\w.,@?^=%&:/~+#-]*[\w@?^=%&/~+#-]\.(m3u8|m3u|mpd))").unwrap();
+    let re = Regex::new(r#"([\"\'])(https?:\/\/[^\"\']*\.(m3u8|m3u|mpd)[^\"\']*)([\"\'])"#).unwrap();
     let links = re
         .captures_iter(text)
-        .map(|caps| caps.get(0).unwrap().as_str().to_string())
-        .collect::<Vec<String>>();
+        .map(|caps| caps.get(2).unwrap().as_str().to_string())
+        .collect::<HashSet<String>>();
 
-    let mut unique_links = vec![];
-    for link in links {
-        if !unique_links.contains(&link) {
-            unique_links.push(link);
-        }
-    }
-    unique_links
+    // in case of amalgated urls
+    links
+        .into_iter()
+        .map(|x| {
+            if x.starts_with("http") {
+                if let Some(y) = x.split("http").last() {
+                    return format!("http{}", y);
+                }
+            }
+            x
+        })
+        .collect()
 }
 
 pub(super) fn decode_base64<T: AsRef<[u8]>>(input: T) -> Result<Vec<u8>> {
@@ -76,10 +83,7 @@ pub(super) fn decrypt_aes_128_cbc(
     let key_length = key.len();
 
     if key_length != 16 {
-        bail!(
-            "invalid key size i.e. {} but expected size 16.",
-            key_length
-        );
+        bail!("invalid key size i.e. {} but expected size 16.", key_length);
     }
 
     let mut key_c = [0_u8; 16];