+++ vsd/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `save`
+  - Handle `--output` flag correctly. ([#21](https://github.com/clitic/vsd/issues/21))
+
 ## [0.3.0] - 2023-08-18
 
 ### Added