+++ vsd/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `save`
+  - `--no-merge` flag. ([#17](https://github.com/clitic/vsd/issues/17), [#20](https://github.com/clitic/vsd/issues/20))
+
 ### Fixed
 
 - `save`