+++ vsd/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Removed self update checker.
+
 ## [0.3.1] - 2024-06-22
 
 ### Added