+++ vsd/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2024-06-23
+
 ### Changed
 
 - Removed self update checker.
@@ -132,6 +134,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2022-06-22
 
 [Unreleased]: https://github.com/clitic/vsd/compare/0.3.1...HEAD
+[0.3.2]: https://github.com/clitic/vsd/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/clitic/vsd/compare/v0.3.0...0.3.1
 [0.3.0]: https://github.com/clitic/vsd/compare/v0.2.5...v0.3.0
 [0.2.5]: https://github.com/clitic/vsd/compare/v0.2.0...v0.2.5